fix: deduplicate find_metadata_cbor and pick newest when multiple exist
Stale build artifacts could leave multiple template_metadata.cbor files
in the build output directory, causing "Multiple metadata files found"
errors. Now picks the most recently modified file.

Also deduplicates the function into a single shared copy in publish.rs,
and shows the error reason when metadata is not found in `tari build`.
